dashboard/backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
google-cloud-pubsub
google-cloud-firestore
